created the banner component
