+++ src/components/IntentSelection/Banner/Banner.style.jsx
@@ -1,0 +1,16 @@
+import { colors } from '../../../utils/constants';
+
+export const banner = {
+  position: 'fixed',
+  zIndex: 1,
+  width: '100%',
+  background: 'white',
+  height: 150,
+  verticalAlign: 'center',
+  borderBottom: `1px solid ${colors.gray}`,
+  padding: '0 15%',
+  marginTop: 50,
+  paddingTop: 80,
+};
+
+export default { banner };
