+++ src/components/IntentSelection/Banner/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './Banner';
